move fixtures up a level to avoid having them run as tests
